Update agent tests for tagged and untagged responses
