+++ .github/workflows/Agents.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        test: [agents-dms]
+        test: [agents-dms, agents-group-tagged, agents-group-untagged]
         environment: [dev, production]
     env:
       DATADOG_API_KEY: ${{ secrets.DATADOG_API_KEY }}
